feat(152787): Distribui o conftest para de acordo com as pastas
